+++ inc/import_memory.class.php
@@ -109,9 +109,12 @@ class PluginFusinvinventoryImport_Memory extends CommonDBTM {
 
 
    
-   function deleteItem($items_id) {
+   function deleteItem($items_id, $idmachine) {
       $CompDevice = new Computer_Device('DeviceMemory');
-      $CompDevice->delete(array("id" => $items_id));
+      $CompDevice->getFromDB($items_id);
+      if ($CompDevice->fields['computers_id'] == $idmachine) {
+         $CompDevice->delete(array("id" => $items_id));
+      }
    }
 
 }
