+++ inc/import_printer.class.php
@@ -106,8 +106,11 @@ class PluginFusinvinventoryImport_Printer extends CommonDBTM {
 
    function deleteItem($items_id) {
       $Computer_Item = new Computer_Item();
-      $Computer_Item->delete(array("id" => $items_id,
-                                   "itemtype" => "Printer"));
+      $Computer_Item->getFromDB($items_id);
+      if ($Computer_Item->fields['computers_id'] == $idmachine) {
+         $Computer_Item->delete(array("id" => $items_id,
+                                      "itemtype" => "Printer"));
+      }
    }
 
 }
