+++ inc/import_processor.class.php
@@ -100,9 +100,12 @@ class PluginFusinvinventoryImport_Processor extends CommonDBTM {
 
 
    
-   function deleteItem($items_id) {
+   function deleteItem($items_id, $idmachine) {
       $CompDevice = new Computer_Device('DeviceProcessor');
-      $CompDevice->delete(array("id" => $items_id));
+      $CompDevice->getFromDB($items_id);
+      if ($CompDevice->fields['computers_id'] == $idmachine) {
+         $CompDevice->delete(array("id" => $items_id));
+      }
    }
 
 }
