+++ inc/import_sound.class.php
@@ -103,9 +103,12 @@ class PluginFusinvinventoryImport_Sound extends CommonDBTM {
 
 
    
-   function deleteItem($items_id) {
+   function deleteItem($items_id, $idmachine) {
       $CompDevice = new Computer_Device('DeviceSoundCard');
-      $CompDevice->delete(array("id" => $items_id));
+      $CompDevice->getFromDB($items_id);
+      if ($CompDevice->fields['computers_id'] == $idmachine) {
+         $CompDevice->delete(array("id" => $items_id));
+      }
    }
 
 }
